Added capability to pass pre-generated UUID to make_ext4fs.
A UUID generated by an external tool, such as uuidgen, can now be
passed to make_ext4fs. It will use that UUID rather than generate its
own. This is helpful if the UUID of the resulting file system must be
known before the file system itself is being created (e.g. filling in
the correct UUID value in the image's own /etc/fstab to mount the file
system on the target).

  $ ./make_ext4fs ... -U 05aceeb5-5827-49cf-b922-10d97cae3c44 ...